paddock-ai-backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone
python-dotenv
requests
beautifulsoup4
lxml 